refactor(mini-bar): extract session grouping into sessionGroups util
